Avoid ungreedy pcre matching
This was using ungreedy matching to remove bits from the html but that
has some built in limitations.  The php pcre docs state it is usually
not possible to match more than pcre.backtrack_limit characters in
ungreedy mode. In practice we've seen this have problems when the
content is many megabytes of wikitext tables, but it can probably
occur elsewhere.

In both use cases here we didn't really need regexes, they happened
to be an easy way to implement what was needed. Switch these use
cases over to simple string manipulations that don't have inherit
length limits.

Bug: T345319
Change-Id: I3c16f3c54d38792a07fcc9f396b5c005e9b5ac63